requirements.txt
# Core web framework
streamlit>=1.28.0

# LangChain ecosystem
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
langchain-huggingface>=0.0.2
langchain-deepseek>=0.0.1
langchain-classic>=0.0.1

# Vector database and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Document processing
pypdf>=3.17.0

# Environment management
python-dotenv>=1.0.0

# Deep learning framework (for sentence-transformers)
torch>=2.0.0
transformers>=4.35.0